fix: pulled out companies from home component
